requirements.txt
google.generativeai
python-dateutil